only one file for lobby tests
